+++ js/pdf-import.js
@@ -1,11 +1,11 @@
 // pdf-import.js — PDF parsing pipeline, import preview, drop zone, batch import
 
 import { state } from './state.js';
-import { MARKER_SCHEMA, SPECIALTY_MARKER_DEFS, calculateCost, formatCost, trackUsage } from './schema.js';
+import { MARKER_SCHEMA, SPECIALTY_MARKER_DEFS, calculateCost, trackUsage } from './schema.js';
 import { IMPORT_STEPS } from './constants.js';
 import { escapeHTML, showNotification, isDebugMode, isPIIReviewEnabled, hashString, showPromptDialog } from './utils.js';
 import { saveImportedData, recalculateHOMAIR } from './data.js';
-import { callClaudeAPI, hasAIProvider, getAIProvider, setAIProvider, setVeniceModel, setOpenRouterModel, getOllamaMainModel, setOllamaMainModel, getVeniceModelDisplay, getOpenRouterModelDisplay, getActiveModelId, getActiveModelDisplay, getOllamaPIIModel, setCustomApiModel, setPpqModel, setRoutstrModel, AI_IMPORT_REQUEST_TIMEOUT_MS } from './api.js';
+import { callClaudeAPI, hasAIProvider, getAIProvider, setAIProvider, setVeniceModel, setOpenRouterModel, setOllamaMainModel, getActiveModelId, setCustomApiModel, setPpqModel, setRoutstrModel, AI_IMPORT_REQUEST_TIMEOUT_MS } from './api.js';
 import { obfuscatePDFText, sanitizeWithOllama, sanitizeWithOllamaStreaming, checkOllamaPII, reviewPIIBeforeSend } from './pii.js';
 import { detectProduct, normalizeWithAdapter, getAdapterByTestType } from './adapters.js';
 import { getPdfDocument } from './pdfjs-loader.js';
@@ -16,8 +16,33 @@ import {
   buildMarkerReference, getExistingImportMarkerKeys,
   normalizeToSI, reconcileImportMarkerMappings,
 } from './pdf-import-marker-mapping.js';
+import {
+  showImportPreview,
+  applyManualImportDate,
+  mapUnmatchedMarker,
+  mapUnmatchedMarkerInput,
+  setImportReviewFilter,
+  applyImportReviewFilters,
+  toggleImportRow,
+  closeImportModal,
+  showImportPreviewAsync,
+  getPendingImport,
+  getExcludedImportIndices,
+  resolveImportPreviewBatch,
+} from './pdf-import-review.js';
 
 export { buildMarkerReference, reconcileImportMarkerMappings } from './pdf-import-marker-mapping.js';
+export {
+  showImportPreview,
+  applyManualImportDate,
+  mapUnmatchedMarker,
+  mapUnmatchedMarkerInput,
+  setImportReviewFilter,
+  applyImportReviewFilters,
+  toggleImportRow,
+  closeImportModal,
+  showImportPreviewAsync,
+} from './pdf-import-review.js';
 
 // ═══════════════════════════════════════════════
 // AI-NEEDED DIALOG — contextual fallback when import is invoked without an AI provider.
@@ -636,335 +661,6 @@ Return ONLY valid JSON in this exact format, no other text:
   };
 }
 
-// ═══════════════════════════════════════════════
-// IMPORT PREVIEW & CONFIRM
-// ═══════════════════════════════════════════════
-export function showImportPreview(parseResult) {
-  const { date, markers, fileName } = parseResult;
-  const modal = document.getElementById("import-modal");
-  const overlay = document.getElementById("import-modal-overlay");
-  const matched = markers.filter(m => m.matched);
-  const newMarkers = markers.filter(m => !m.matched && m.suggestedKey);
-  const unmatched = markers.filter(m => !m.matched && !m.suggestedKey);
-  const importCount = matched.length + newMarkers.length;
-  const batchCtx = window._batchImportContext;
-  const batchLabel = batchCtx ? `File ${batchCtx.current} of ${batchCtx.total}` : 'Lab import';
-  modal.className = 'modal import-preview-modal';
-  let html = `<div class="gb-modal-head import-preview-head">
-    <div>
-      <div class="gb-modal-kicker">${escapeHTML(batchLabel)}</div>
-      <div class="gb-modal-title">Review &amp; Edit Import</div>
-    </div>
-    <button type="button" class="modal-close" onclick="closeImportModal()" aria-label="Close import review">&times;</button>
-  </div>
-  <div class="gb-form-body import-review-body">
-    <div class="import-review-summary">
-      <div class="import-review-file">
-        <span class="import-review-label">File</span>
-        <strong>${escapeHTML(fileName)}</strong>
-      </div>
-      <div class="import-review-file">
-        <span class="import-review-label">Collection date</span>
-        <input type="date" id="import-manual-date" value="${escapeHTML(date || '')}" onchange="applyManualImportDate(this.value)" aria-label="Collection date">
-      </div>
-      <div class="import-review-stats" aria-label="Import mapping summary">
-        <span class="import-review-stat import-review-stat-matched"><strong>${matched.length}</strong> matched</span>
-        <span class="import-review-stat import-review-stat-new"><strong>${newMarkers.length}</strong> new</span>
-        <span class="import-review-stat import-review-stat-unmatched"><strong>${unmatched.length}</strong> unmatched</span>
-      </div>
-    </div>`;
-  // Quality warning — high unmatched ratio suggests unsupported lab
-  const unmatchedRatio = markers.length > 0 ? unmatched.length / markers.length : 0;
-  if (unmatchedRatio > 0.4 && unmatched.length > 10) {
-    html += `<div class="import-review-warning">
-      A large portion of markers couldn't be mapped. This lab report may not be well supported yet — review the results below carefully before importing.
-      You can <a href="https://github.com/elkimek/get-based/issues" target="_blank" rel="noopener">request support</a> for this lab on GitHub.</div>`;
-  }
-  if (!date) {
-    html += `<div class="import-review-warning import-review-date-warning">
-      Could not extract collection date from PDF. Please set it above before importing.</div>`;
-  }
-  // Build reference lookup (used for unmatched dropdown + range comparison)
-  const refLookup = buildMarkerReference();
-  const allKeys = Object.entries(refLookup).map(([key, def]) => ({ key, name: def.name }));
-  allKeys.sort((a, b) => a.name.localeCompare(b.name));
-  const optionsHtml = allKeys.map(k => {
-    const label = `${k.name} (${k.key})`;
-    return `<option value="${escapeHTML(k.key)}" label="${escapeHTML(label)}"></option>`;
-  }).join('');
-
-  html += `<div class="import-review-controls">
-    <div class="import-filter-group" role="group" aria-label="Filter import rows">
-      <button type="button" class="import-filter-btn active" data-filter="all" onclick="setImportReviewFilter(this)">All</button>
-      <button type="button" class="import-filter-btn" data-filter="matched" onclick="setImportReviewFilter(this)">Matched</button>
-      <button type="button" class="import-filter-btn" data-filter="new" onclick="setImportReviewFilter(this)">New</button>
-      <button type="button" class="import-filter-btn" data-filter="unmatched" onclick="setImportReviewFilter(this)">Unmatched</button>
-      <button type="button" class="import-filter-btn" data-filter="excluded" onclick="setImportReviewFilter(this)">Excluded</button>
-    </div>
-    <label class="import-review-search-wrap">
-      <span class="sr-only">Search import rows</span>
-      <input type="search" id="import-review-search" class="import-review-search" placeholder="Search markers" oninput="applyImportReviewFilters()" autocomplete="off">
-    </label>
-    <span class="import-visible-count" id="import-visible-count" aria-live="polite"></span>
-  </div>`;
-
-  html += `<div class="import-table-wrap"><table class="import-table"><thead><tr><th>Status</th><th>Test Name</th><th>Value</th><th>Lab Range</th><th>Maps To</th><th>Action</th></tr></thead><tbody>`;
-  for (const m of matched) {
-    const origIdx = markers.indexOf(m);
-    const labRange = (m.refMin != null || m.refMax != null) ? `${m.refMin ?? '?'}\u2013${m.refMax ?? '?'}` : '';
-    html += `<tr data-import-idx="${origIdx}" data-import-status="matched">
-      <td class="import-status-cell matched" data-label="Status"><span class="import-status-pill">Matched</span></td>
-      <td class="import-name-cell" data-label="Test name">${escapeHTML(m.rawName)}</td>
-      <td data-label="Value">${escapeHTML(String(m.value))}</td>
-      <td class="import-range-cell" data-label="Lab range">${escapeHTML(labRange || '—')}</td>
-      <td class="import-map-cell" data-label="Maps to">${escapeHTML(m.mappedKey)}</td>
-      <td class="import-row-action" data-label="Action"><button type="button" class="import-exclude-btn" onclick="toggleImportRow(this)" title="Exclude from import" aria-label="Exclude ${escapeHTML(m.rawName)} from import">Exclude</button></td>
-    </tr>`;
-  }
-  for (const m of newMarkers) {
-    const origIdx = markers.indexOf(m);
-    const labRange = (m.refMin != null || m.refMax != null) ? `${m.refMin ?? '?'}\u2013${m.refMax ?? '?'}` : '';
-    html += `<tr data-import-idx="${origIdx}" data-import-status="new">
-      <td class="import-status-cell new-marker" data-label="Status"><span class="import-status-pill">New</span></td>
-      <td class="import-name-cell" data-label="Test name">${escapeHTML(m.rawName)}</td>
-      <td data-label="Value">${escapeHTML(String(m.value))}</td>
-      <td class="import-range-cell" data-label="Lab range">${escapeHTML(labRange || '—')}</td>
-      <td class="import-map-cell" data-label="Maps to">${escapeHTML(m.suggestedKey)}</td>
-      <td class="import-row-action" data-label="Action"><button type="button" class="import-exclude-btn" onclick="toggleImportRow(this)" title="Exclude from import" aria-label="Exclude ${escapeHTML(m.rawName)} from import">Exclude</button></td>
-    </tr>`;
-  }
-  if (unmatched.length > 0) {
-    for (const m of unmatched) {
-      const origIdx = markers.indexOf(m);
-      const labRange = (m.refMin != null || m.refMax != null) ? `${m.refMin ?? '?'}\u2013${m.refMax ?? '?'}` : '';
-      html += `<tr data-import-idx="${origIdx}" data-import-status="unmatched">
-        <td class="import-status-cell unmatched" data-label="Status"><span class="import-status-pill">Unmatched</span></td>
-        <td class="import-name-cell" data-label="Test name">${escapeHTML(m.rawName)}</td>
-        <td data-label="Value">${escapeHTML(String(m.value))}</td>
-        <td class="import-range-cell" data-label="Lab range">${escapeHTML(labRange || '—')}</td>
-        <td class="import-map-cell" data-label="Maps to">
-          <input type="text" class="import-map-input" list="import-marker-options" data-marker-idx="${origIdx}" onchange="mapUnmatchedMarkerInput(this)" placeholder="Search marker" autocomplete="off" aria-label="Map ${escapeHTML(m.rawName)} to an existing marker">
-        </td>
-        <td class="import-row-action" data-label="Action"><span class="import-skip-note">Skipped unless mapped</span></td>
-      </tr>`;
-    }
-  }
-  html += `</tbody></table></div>`;
-  if (unmatched.length > 0) {
-    html += `<datalist id="import-marker-options">${optionsHtml}</datalist>`;
-  }
-
-  // Reference range adoption toggle — count matched markers where PDF ranges differ from current
-  let rangesDiffCount = 0;
-  for (const m of matched) {
-    if (m.refMin == null && m.refMax == null) continue;
-    const schemaRef = refLookup[m.mappedKey];
-    if (!schemaRef) continue;
-    // Normalize PDF ranges to SI for accurate comparison against schema
-    const siMin = m.refMin != null ? normalizeToSI(m.mappedKey, m.refMin, m.unit) : null;
-    const siMax = m.refMax != null ? normalizeToSI(m.mappedKey, m.refMax, m.unit) : null;
-    if ((siMin !== schemaRef.refMin && !(siMin != null && schemaRef.refMin != null && Math.abs(siMin - schemaRef.refMin) < 0.001)) ||
-        (siMax !== schemaRef.refMax && !(siMax != null && schemaRef.refMax != null && Math.abs(siMax - schemaRef.refMax) < 0.001))) {
-      rangesDiffCount++;
-    }
-  }
-  if (rangesDiffCount > 0) {
-    html += `<label class="import-range-option">
-      <input type="checkbox" id="import-adopt-ranges">
-      <span><strong>Update reference ranges from this report</strong><small>${rangesDiffCount} marker${rangesDiffCount !== 1 ? 's' : ''} differ from the current ranges. Leave off unless you want this lab's ranges to become the active reference.</small></span></label>`;
-  }
-
-  // Privacy notice
-  if (parseResult.privacyMethod?.startsWith('ollama')) {
-    html += `<div class="privacy-notice privacy-notice-success">&#128274; Personal information scrubbed by local AI${parseResult.privacyMethod === 'ollama+review' ? ' (reviewed)' : ''}</div>`;
-  } else if (parseResult.privacyMethod === 'regex') {
-    html += `<div class="privacy-notice privacy-notice-warning">&#128274; ${parseResult.privacyReplacements} personal detail${parseResult.privacyReplacements !== 1 ? 's' : ''} replaced with fake data`;
-    html += `<span class="privacy-notice-detail">Set up Local AI in Settings for comprehensive language-aware protection</span></div>`;
-  }
-  // Cost info (always visible)
-  if (parseResult.costInfo) {
-    const ci = parseResult.costInfo;
-    const totalTokens = (ci.inputTokens || 0) + (ci.outputTokens || 0);
-    const modelLabel = ci.provider === 'ollama' ? getOllamaMainModel() : ci.provider === 'venice' ? getVeniceModelDisplay() : ci.provider === 'openrouter' ? getOpenRouterModelDisplay() : getActiveModelDisplay();
-    html += `<div class="import-cost-note">\ud83d\udcca ${escapeHTML(modelLabel)} \u00b7 ${totalTokens.toLocaleString()} tokens \u00b7 ${formatCost(ci.cost)}</div>`;
-  }
-  // Debug: timings and diff button
-  if (isDebugMode()) {
-    const t = parseResult.timings;
-    if (t) {
-      const piiLabel = parseResult.privacyMethod?.startsWith('ollama') ? `PII: ${t.pii}s (${getOllamaPIIModel()})` : `PII: regex`;
-      const provider = getAIProvider();
-      const modelLabel = provider === 'ollama' ? getOllamaMainModel() : provider === 'venice' ? getVeniceModelDisplay() : provider === 'openrouter' ? getOpenRouterModelDisplay() : getActiveModelDisplay();
-      html += `<div class="import-debug-note">&#9202; ${piiLabel} &nbsp;|&nbsp; Analysis: ${t.analysis}s (${modelLabel})</div>`;
-    }
-    if (parseResult.privacyOriginal && parseResult.privacyObfuscated) {
-      html += `<button type="button" class="import-btn import-btn-secondary import-privacy-details-btn" onclick="showPIIDiffViewer(window._pendingImport.privacyOriginal, window._pendingImport.privacyObfuscated)">&#128269; View privacy details</button>`;
-    }
-  }
-
-  const cancelLabel = batchCtx ? 'Skip' : 'Cancel';
-  const importDisabled = !date ? ' disabled' : '';
-  html += `</div>
-    <div class="import-review-actions">
-      <button type="button" class="import-btn import-btn-secondary" onclick="closeImportModal()">${cancelLabel}</button>
-      <button type="button" class="import-btn import-btn-primary" id="import-confirm-btn" onclick="confirmImport()"${importDisabled}>Import ${importCount} Marker${importCount !== 1 ? 's' : ''}</button>
-    </div>`;
-  if (!parseResult._importProfileId) parseResult._importProfileId = state.currentProfile;
-  window._pendingImport = parseResult;
-  window._pendingImportRefLookup = refLookup;
-  modal.innerHTML = html;
-  overlay.classList.add("show");
-  applyImportReviewFilters();
-}
-
-export function mapUnmatchedMarker(selectEl) {
-  applyImportMarkerMapping(selectEl, selectEl.value || '');
-}
-
-export function mapUnmatchedMarkerInput(inputEl) {
-  const raw = inputEl.value.trim();
-  const key = resolveImportMarkerKey(raw);
-  if (raw && !key) {
-    inputEl.value = '';
-    showNotification('Choose a marker from the list', 'error');
-    applyImportMarkerMapping(inputEl, '');
-    return;
-  }
-  inputEl.value = key || '';
-  applyImportMarkerMapping(inputEl, key || '');
-}
-
-function resolveImportMarkerKey(raw) {
-  if (!raw) return '';
-  const refLookup = window._pendingImportRefLookup || buildMarkerReference();
-  if (refLookup[raw]) return raw;
-  const normalized = raw.toLowerCase();
-  for (const [key, def] of Object.entries(refLookup)) {
-    const name = String(def.name || '').toLowerCase();
-    if (key.toLowerCase() === normalized || name === normalized || `${name} (${key.toLowerCase()})` === normalized) {
-      return key;
-    }
-  }
-  return '';
-}
-
-function applyImportMarkerMapping(controlEl, key) {
-  const result = window._pendingImport;
-  if (!result) return;
-  const idx = parseInt(controlEl.dataset.markerIdx, 10);
-  const marker = result.markers[idx];
-  if (!marker) return;
-  marker.mappedKey = key || null;
-  marker.matched = !!key;
-  const row = controlEl.closest('tr');
-  if (row) {
-    const statusCell = row.querySelector('td:first-child');
-    const actionCell = row.querySelector('.import-row-action');
-    if (key) {
-      row.dataset.importStatus = 'matched';
-      if (statusCell) {
-        statusCell.className = 'import-status-cell matched';
-        statusCell.innerHTML = '<span class="import-status-pill">Matched</span>';
-      }
-      if (actionCell && !actionCell.querySelector('.import-exclude-btn')) {
-        actionCell.innerHTML = '<button type="button" class="import-exclude-btn" onclick="toggleImportRow(this)" title="Exclude from import" aria-label="Exclude from import">Exclude</button>';
-      }
-    } else {
-      row.dataset.importStatus = 'unmatched';
-      row.classList.remove('import-excluded');
-      if (statusCell) {
-        statusCell.className = 'import-status-cell unmatched';
-        statusCell.innerHTML = '<span class="import-status-pill">Unmatched</span>';
-      }
-      if (actionCell) actionCell.innerHTML = '<span class="import-skip-note">Skipped unless mapped</span>';
-    }
-  }
-  updateImportConfirmCount();
-  applyImportReviewFilters();
-}
-
-function updateImportConfirmCount() {
-  const result = window._pendingImport;
-  if (!result) return;
-  const excludedIdxs = _getExcludedIndices();
-  const importCount = result.markers.filter((m, i) => (m.matched || (!m.matched && m.suggestedKey)) && !excludedIdxs.has(i)).length;
-  const btn = document.getElementById('import-confirm-btn');
-  if (btn) btn.textContent = `Import ${importCount} Marker${importCount !== 1 ? 's' : ''}`;
-}
-
-export function setImportReviewFilter(btn) {
-  const group = btn.closest('.import-filter-group');
-  if (group) {
-    for (const item of group.querySelectorAll('.import-filter-btn')) item.classList.toggle('active', item === btn);
-  }
-  applyImportReviewFilters();
-}
-
-export function applyImportReviewFilters() {
-  const rows = Array.from(document.querySelectorAll('.import-table tbody tr[data-import-idx]'));
-  if (rows.length === 0) return;
-  const activeFilter = document.querySelector('.import-filter-btn.active')?.dataset.filter || 'all';
-  const query = (document.getElementById('import-review-search')?.value || '').trim().toLowerCase();
-  let visible = 0;
-  for (const row of rows) {
-    const status = row.classList.contains('import-excluded') ? 'excluded' : (row.dataset.importStatus || '');
-    const filterMatch = activeFilter === 'all' || activeFilter === status;
-    const controlText = Array.from(row.querySelectorAll('input, select')).map(el => el.value).join(' ');
-    const searchMatch = !query || `${row.textContent} ${controlText}`.toLowerCase().includes(query);
-    const shouldShow = filterMatch && searchMatch;
-    row.hidden = !shouldShow;
-    if (shouldShow) visible++;
-  }
-  const count = document.getElementById('import-visible-count');
-  if (count) count.textContent = `${visible}/${rows.length} shown`;
-}
-
-export function applyManualImportDate(dateStr) {
-  const btn = document.getElementById('import-confirm-btn');
-  if (!window._pendingImport) return;
-  const nextDate = (dateStr || '').trim();
-  window._pendingImport.date = nextDate;
-  if (btn) {
-    btn.disabled = !nextDate;
-    btn.style.opacity = '';
-    btn.style.cursor = '';
-  }
-}
-
-export function toggleImportRow(btn) {
-  const row = btn.closest('tr');
-  if (!row) return;
-  const excluded = row.classList.toggle('import-excluded');
-  btn.textContent = excluded ? 'Include' : 'Exclude';
-  btn.title = excluded ? 'Include in import' : 'Exclude from import';
-  btn.setAttribute('aria-label', btn.title);
-  updateImportConfirmCount();
-  applyImportReviewFilters();
-}
-
-function _getExcludedIndices() {
-  const excluded = new Set();
-  for (const row of document.querySelectorAll('.import-table tr.import-excluded[data-import-idx]')) {
-    excluded.add(parseInt(row.dataset.importIdx, 10));
-  }
-  return excluded;
-}
-
-export function closeImportModal() {
-  document.getElementById("import-modal-overlay").classList.remove("show");
-  window._pendingImport = null;
-  window._pendingImportRefLookup = null;
-  // Restore batch progress visibility for the next file
-  const dropZone = document.getElementById("drop-zone");
-  if (dropZone) dropZone.style.display = '';
-  if (window._batchImportResolve) {
-    const resolve = window._batchImportResolve;
-    window._batchImportResolve = null;
-    window._batchImportContext = null;
-    resolve('skip');
-  }
-}
-
 function snapshotImportedData() {
   try { return JSON.stringify(state.importedData || {}); } catch { return null; }
 }
@@ -984,25 +680,25 @@ function isValidISOCalendarDate(date) {
 }
 
 export async function confirmImport() {
-  const result = window._pendingImport;
+  const result = getPendingImport();
   if (!result || !result.date) return;
   const confirmBtn = document.getElementById('import-confirm-btn');
   if (confirmBtn) confirmBtn.disabled = true;
   // Guard: if profile changed during async import, abort to prevent saving to wrong profile
   if (result._importProfileId && result._importProfileId !== state.currentProfile) {
     showNotification('Profile changed during import — import cancelled for safety.', 'error');
-    window.closeImportModal();
+    closeImportModal();
     return;
   }
   const rollback = snapshotImportedData();
-  const excludedIdxs = _getExcludedIndices();
+  const excludedIdxs = getExcludedImportIndices();
   const matched = result.markers.filter((m, i) => m.matched && !excludedIdxs.has(i));
   const newMarkers = result.markers.filter((m, i) => !m.matched && m.suggestedKey && !excludedIdxs.has(i));
   const importCount = matched.length + newMarkers.length;
   if (importCount === 0) {
     showNotification("No markers to import", "error");
     if (confirmBtn) confirmBtn.disabled = false;
-    window.closeImportModal();
+    closeImportModal();
     return;
   }
   if (!state.importedData.entries) state.importedData.entries = [];
@@ -1113,20 +809,8 @@ export async function confirmImport() {
     if (confirmBtn) confirmBtn.disabled = false;
     return;
   }
-  // Resolve batch promise before closeImportModal (which would resolve with 'skip')
-  if (window._batchImportResolve) {
-    const resolve = window._batchImportResolve;
-    window._batchImportResolve = null;
-    window._batchImportContext = null;
-    document.getElementById("import-modal-overlay").classList.remove("show");
-    window._pendingImport = null;
-    // Restore batch progress visibility for the next file
-    const dropZone = document.getElementById("drop-zone");
-    if (dropZone) dropZone.style.display = '';
-    resolve('import');
-  } else {
-    window.closeImportModal();
-  }
+  // Resolve batch promise before closeImportModal (which would resolve with 'skip').
+  if (!resolveImportPreviewBatch('import')) closeImportModal();
   // During batch mode, defer expensive UI refreshes until the batch completes
   if (!_batchMode) {
     window.buildSidebar();
@@ -2008,17 +1692,6 @@ export async function showBatchImportProgress(step, fileName, current, total) {
   dropZone.innerHTML = html;
   _observeProgressBar();
   await new Promise(r => requestAnimationFrame(() => setTimeout(r, 0)));
-}
-
-export function showImportPreviewAsync(result, fileName, current, total) {
-  // Hide batch progress so it doesn't cover the import preview modal
-  const dropZone = document.getElementById("drop-zone");
-  if (dropZone) dropZone.style.display = 'none';
-  return new Promise(resolve => {
-    window._batchImportResolve = resolve;
-    window._batchImportContext = { current, total };
-    showImportPreview(result);
-  });
 }
 
 // ═══════════════════════════════════════════════
