+++ service-worker.js
@@ -106,6 +106,7 @@ const APP_SHELL = [
   '/js/emf.js',
   '/js/image-utils.js',
   '/js/pdf-import.js',
+  '/js/pdf-import-review.js',
   '/js/pdf-import-marker-mapping.js',
   '/js/export.js',
   '/js/chat.js',
